+++ app/resources/numismatics/coins/numismatics/coins_controller.rb
@@ -11,13 +11,11 @@ module Numismatics
     )
 
     before_action :load_facet_values, only: [:new, :edit]
-    before_action :load_holding_locations, only: [:new, :edit]
     before_action :load_numismatic_collections, only: [:new, :edit]
     before_action :selected_issue, only: [:new, :edit, :destroy]
 
     def facet_fields
       [
-        :holding_location_ssim,
         :numismatic_collection_ssim
       ]
     end
@@ -50,10 +48,6 @@ module Numismatics
     def load_facet_values
       query = FindFacetValues.new(query_service: Valkyrie::MetadataAdapter.find(:index_solr).query_service)
       @facet_values = query.find_facet_values(facet_fields: facet_fields)
-    end
-
-    def load_holding_locations
-      @holding_locations = @facet_values[:holding_location_ssim]
     end
 
     def load_numismatic_collections